move TestUi func to sdk
