+++ builder/azure/chroot/step_create_shared_image_version_test.go
@@ -9,8 +9,8 @@ import (
 	"testing"
 
 	"github.com/hashicorp/packer/builder/azure/common/client"
-	"github.com/hashicorp/packer/packer"
 	"github.com/hashicorp/packer/packer-plugin-sdk/multistep"
+	packersdk "github.com/hashicorp/packer/packer-plugin-sdk/packer"
 
 	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2019-12-01/compute"
 	"github.com/Azure/go-autorest/autorest"
@@ -126,7 +126,7 @@ func TestStepCreateSharedImageVersion_Run(t *testing.T) {
 		state.Put("azureclient", &client.AzureClientSetMock{
 			GalleryImageVersionsClientMock: m,
 		})
-		state.Put("ui", packer.TestUi(t))
+		state.Put("ui", packersdk.TestUi(t))
 		state.Put(stateBagKey_Snapshotset, tt.snapshotset)
 
 		t.Run(tt.name, func(t *testing.T) {
