+++ builder/docker/communicator_test.go
@@ -18,7 +18,7 @@ import (
 
 // TestUploadDownload verifies that basic upload / download functionality works
 func TestUploadDownload(t *testing.T) {
-	ui := packer.TestUi(t)
+	ui := packersdk.TestUi(t)
 
 	tpl, err := template.Parse(strings.NewReader(dockerBuilderConfig))
 	if err != nil {
@@ -101,7 +101,7 @@ func TestUploadDownload(t *testing.T) {
 // need to use github.com/cbednarski/rerun to verify since this problem occurs
 // only intermittently.
 func TestLargeDownload(t *testing.T) {
-	ui := packer.TestUi(t)
+	ui := packersdk.TestUi(t)
 
 	tpl, err := template.Parse(strings.NewReader(dockerLargeBuilderConfig))
 	if err != nil {
@@ -205,7 +205,7 @@ func TestLargeDownload(t *testing.T) {
 
 // TestFixUploadOwner verifies that owner of uploaded files is the user the container is running as.
 func TestFixUploadOwner(t *testing.T) {
-	ui := packer.TestUi(t)
+	ui := packersdk.TestUi(t)
 
 	tpl, err := template.Parse(strings.NewReader(testFixUploadOwnerTemplate))
 	if err != nil {
