+++ builder/proxmox/common/step_convert_to_template_test.go
@@ -6,8 +6,8 @@ import (
 	"testing"
 
 	"github.com/Telmate/proxmox-api-go/proxmox"
-	"github.com/hashicorp/packer/packer"
 	"github.com/hashicorp/packer/packer-plugin-sdk/multistep"
+	packersdk "github.com/hashicorp/packer/packer-plugin-sdk/packer"
 )
 
 type converterMock struct {
@@ -79,7 +79,7 @@ func TestConvertToTemplate(t *testing.T) {
 			}
 
 			state := new(multistep.BasicStateBag)
-			state.Put("ui", packer.TestUi(t))
+			state.Put("ui", packersdk.TestUi(t))
 			state.Put("vmRef", proxmox.NewVmRef(vmid))
 			state.Put("proxmoxClient", converter)
