+++ builder/proxmox/common/step_type_boot_command_test.go
@@ -7,9 +7,9 @@ import (
 	"testing"
 
 	"github.com/Telmate/proxmox-api-go/proxmox"
-	"github.com/hashicorp/packer/packer"
 	"github.com/hashicorp/packer/packer-plugin-sdk/bootcommand"
 	"github.com/hashicorp/packer/packer-plugin-sdk/multistep"
+	packersdk "github.com/hashicorp/packer/packer-plugin-sdk/packer"
 )
 
 type commandTyperMock struct {
@@ -128,7 +128,7 @@ func TestTypeBootCommand(t *testing.T) {
 			}
 
 			state := new(multistep.BasicStateBag)
-			state.Put("ui", packer.TestUi(t))
+			state.Put("ui", packersdk.TestUi(t))
 			state.Put("config", c.builderConfig)
 			state.Put("http_port", int(0))
 			state.Put("vmRef", proxmox.NewVmRef(1))
