+++ builder/qemu/step_run_test.go
@@ -4,10 +4,10 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/hashicorp/packer/packer"
 	"github.com/hashicorp/packer/packer-plugin-sdk/communicator"
 	"github.com/hashicorp/packer/packer-plugin-sdk/multistep"
 	"github.com/hashicorp/packer/packer-plugin-sdk/multistep/commonsteps"
+	packersdk "github.com/hashicorp/packer/packer-plugin-sdk/packer"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -92,7 +92,7 @@ func Test_UserOverrides(t *testing.T) {
 
 		step := &stepRun{
 			atLeastVersion2: true,
-			ui:              packer.TestUi(t),
+			ui:              packersdk.TestUi(t),
 		}
 		args, err := step.getCommandArgs(tc.Config, state)
 		if err != nil {
@@ -130,7 +130,7 @@ func Test_DriveAndDeviceArgs(t *testing.T) {
 			map[string]interface{}{},
 			&stepRun{
 				atLeastVersion2: true,
-				ui:              packer.TestUi(t),
+				ui:              packersdk.TestUi(t),
 			},
 			[]string{
 				"-display", "gtk",
@@ -155,7 +155,7 @@ func Test_DriveAndDeviceArgs(t *testing.T) {
 			&stepRun{
 				DiskImage:       true,
 				atLeastVersion2: true,
-				ui:              packer.TestUi(t),
+				ui:              packersdk.TestUi(t),
 			},
 			[]string{
 				"-display", "gtk",
@@ -183,7 +183,7 @@ func Test_DriveAndDeviceArgs(t *testing.T) {
 			&stepRun{
 				DiskImage:       true,
 				atLeastVersion2: true,
-				ui:              packer.TestUi(t),
+				ui:              packersdk.TestUi(t),
 			},
 			[]string{
 				"-display", "gtk",
@@ -212,7 +212,7 @@ func Test_DriveAndDeviceArgs(t *testing.T) {
 			},
 			&stepRun{
 				atLeastVersion2: true,
-				ui:              packer.TestUi(t),
+				ui:              packersdk.TestUi(t),
 			},
 			[]string{
 				"-display", "gtk",
@@ -244,7 +244,7 @@ func Test_DriveAndDeviceArgs(t *testing.T) {
 			},
 			&stepRun{
 				atLeastVersion2: true,
-				ui:              packer.TestUi(t),
+				ui:              packersdk.TestUi(t),
 			},
 			[]string{
 				"-display", "gtk",
@@ -274,7 +274,7 @@ func Test_DriveAndDeviceArgs(t *testing.T) {
 			},
 			&stepRun{
 				atLeastVersion2: true,
-				ui:              packer.TestUi(t),
+				ui:              packersdk.TestUi(t),
 			},
 			[]string{
 				"-display", "gtk",
@@ -294,7 +294,7 @@ func Test_DriveAndDeviceArgs(t *testing.T) {
 			},
 			&stepRun{
 				atLeastVersion2: true,
-				ui:              packer.TestUi(t),
+				ui:              packersdk.TestUi(t),
 			},
 			[]string{
 				"-display", "gtk",
@@ -314,7 +314,7 @@ func Test_DriveAndDeviceArgs(t *testing.T) {
 			},
 			&stepRun{
 				atLeastVersion2: true,
-				ui:              packer.TestUi(t),
+				ui:              packersdk.TestUi(t),
 			},
 			[]string{
 				"-display", "gtk",
@@ -338,7 +338,7 @@ func Test_DriveAndDeviceArgs(t *testing.T) {
 			},
 			&stepRun{
 				atLeastVersion2: false,
-				ui:              packer.TestUi(t),
+				ui:              packersdk.TestUi(t),
 			},
 			[]string{
 				"-boot", "once=d",
@@ -360,7 +360,7 @@ func Test_DriveAndDeviceArgs(t *testing.T) {
 			},
 			&stepRun{
 				atLeastVersion2: true,
-				ui:              packer.TestUi(t),
+				ui:              packersdk.TestUi(t),
 			},
 			[]string{
 				"-display", "gtk",
@@ -386,7 +386,7 @@ func Test_DriveAndDeviceArgs(t *testing.T) {
 			&stepRun{
 				DiskImage:       true,
 				atLeastVersion2: true,
-				ui:              packer.TestUi(t),
+				ui:              packersdk.TestUi(t),
 			},
 			[]string{
 				"-display", "gtk",
@@ -436,7 +436,7 @@ func Test_OptionalConfigOptionsGetSet(t *testing.T) {
 	state := runTestState(t, c)
 	step := &stepRun{
 		atLeastVersion2: true,
-		ui:              packer.TestUi(t),
+		ui:              packersdk.TestUi(t),
 	}
 	args, err := step.getCommandArgs(c, state)
 	if err != nil {
@@ -475,7 +475,7 @@ func Test_Defaults(t *testing.T) {
 		{
 			&Config{},
 			map[string]interface{}{},
-			&stepRun{ui: packer.TestUi(t)},
+			&stepRun{ui: packersdk.TestUi(t)},
 			[]string{"-boot", "once=d"},
 			"Boot value should default to once=d",
 		},
@@ -484,7 +484,7 @@ func Test_Defaults(t *testing.T) {
 			map[string]interface{}{},
 			&stepRun{
 				DiskImage: true,
-				ui:        packer.TestUi(t),
+				ui:        packersdk.TestUi(t),
 			},
 			[]string{"-boot", "c"},
 			"Boot value should be set to c when DiskImage is set on step",
@@ -495,7 +495,7 @@ func Test_Defaults(t *testing.T) {
 				QMPSocketPath: "/path/to/socket",
 			},
 			map[string]interface{}{},
-			&stepRun{ui: packer.TestUi(t)},
+			&stepRun{ui: packersdk.TestUi(t)},
 			[]string{"-qmp", "unix:/path/to/socket,server,nowait"},
 			"Args should contain -qmp when qmp_enable is set",
 		},
@@ -504,7 +504,7 @@ func Test_Defaults(t *testing.T) {
 				QMPEnable: true,
 			},
 			map[string]interface{}{},
-			&stepRun{ui: packer.TestUi(t)},
+			&stepRun{ui: packersdk.TestUi(t)},
 			[]string{"-qmp", "unix:,server,nowait"},
 			"Args contain -qmp even when socket path isn't set, if qmp enabled",
 		},
@@ -513,14 +513,14 @@ func Test_Defaults(t *testing.T) {
 				VMName: "partyname",
 			},
 			map[string]interface{}{},
-			&stepRun{ui: packer.TestUi(t)},
+			&stepRun{ui: packersdk.TestUi(t)},
 			[]string{"-name", "partyname"},
 			"Name is set from config",
 		},
 		{
 			&Config{},
 			map[string]interface{}{},
-			&stepRun{ui: packer.TestUi(t)},
+			&stepRun{ui: packersdk.TestUi(t)},
 			[]string{"-name", ""},
 			"Name is set from config, even when name is blank (which won't " +
 				"happen for real thanks to defaulting in build prepare)",
@@ -531,7 +531,7 @@ func Test_Defaults(t *testing.T) {
 				MachineType: "fancymachine",
 			},
 			map[string]interface{}{},
-			&stepRun{ui: packer.TestUi(t)},
+			&stepRun{ui: packersdk.TestUi(t)},
 			[]string{"-machine", "type=fancymachine"},
 			"Don't add accelerator tag when no accelerator is set.",
 		},
@@ -541,7 +541,7 @@ func Test_Defaults(t *testing.T) {
 				MachineType: "fancymachine",
 			},
 			map[string]interface{}{},
-			&stepRun{ui: packer.TestUi(t)},
+			&stepRun{ui: packersdk.TestUi(t)},
 			[]string{"-machine", "type=fancymachine,accel=kvm"},
 			"Add accelerator tag when accelerator is set.",
 		},
@@ -550,7 +550,7 @@ func Test_Defaults(t *testing.T) {
 				NetBridge: "fakebridge",
 			},
 			map[string]interface{}{},
-			&stepRun{ui: packer.TestUi(t)},
+			&stepRun{ui: packersdk.TestUi(t)},
 			[]string{"-netdev", "bridge,id=user.0,br=fakebridge"},
 			"Add netbridge tag when netbridge is set.",
 		},
@@ -563,7 +563,7 @@ func Test_Defaults(t *testing.T) {
 				},
 			},
 			map[string]interface{}{},
-			&stepRun{ui: packer.TestUi(t)},
+			&stepRun{ui: packersdk.TestUi(t)},
 			[]string{"-netdev", "user,id=user.0"},
 			"No host forwarding when no net bridge and no communicator",
 		},
@@ -581,7 +581,7 @@ func Test_Defaults(t *testing.T) {
 			map[string]interface{}{
 				"commHostPort": 1111,
 			},
-			&stepRun{ui: packer.TestUi(t)},
+			&stepRun{ui: packersdk.TestUi(t)},
 			[]string{"-netdev", "user,id=user.0,hostfwd=tcp::1111-:4567"},
 			"Host forwarding when a communicator is configured",
 		},
@@ -592,7 +592,7 @@ func Test_Defaults(t *testing.T) {
 			map[string]interface{}{
 				"vnc_port": 5959,
 			},
-			&stepRun{ui: packer.TestUi(t)},
+			&stepRun{ui: packersdk.TestUi(t)},
 			[]string{"-vnc", "1.1.1.1:5959"},
 			"no VNC password should be set",
 		},
@@ -604,7 +604,7 @@ func Test_Defaults(t *testing.T) {
 			map[string]interface{}{
 				"vnc_port": 5959,
 			},
-			&stepRun{ui: packer.TestUi(t)},
+			&stepRun{ui: packersdk.TestUi(t)},
 			[]string{"-vnc", "1.1.1.1:5959,password"},
 			"VNC password should be set",
 		},
@@ -613,7 +613,7 @@ func Test_Defaults(t *testing.T) {
 				MemorySize: 2345,
 			},
 			map[string]interface{}{},
-			&stepRun{ui: packer.TestUi(t)},
+			&stepRun{ui: packersdk.TestUi(t)},
 			[]string{"-m", "2345M"},
 			"Memory is set, with unit M",
 		},
@@ -622,7 +622,7 @@ func Test_Defaults(t *testing.T) {
 				CpuCount: 2,
 			},
 			map[string]interface{}{},
-			&stepRun{ui: packer.TestUi(t)},
+			&stepRun{ui: packersdk.TestUi(t)},
 			[]string{"-smp", "cpus=2,sockets=2"},
 			"both cpus and sockets are set to config's CpuCount",
 		},
@@ -631,7 +631,7 @@ func Test_Defaults(t *testing.T) {
 				CpuCount: 2,
 			},
 			map[string]interface{}{},
-			&stepRun{ui: packer.TestUi(t)},
+			&stepRun{ui: packersdk.TestUi(t)},
 			[]string{"-smp", "cpus=2,sockets=2"},
 			"both cpus and sockets are set to config's CpuCount",
 		},
@@ -642,7 +642,7 @@ func Test_Defaults(t *testing.T) {
 			map[string]interface{}{
 				"floppy_path": "/path/to/floppy",
 			},
-			&stepRun{ui: packer.TestUi(t)},
+			&stepRun{ui: packersdk.TestUi(t)},
 			[]string{"-fda", "/path/to/floppy"},
 			"floppy path should be set under fda flag, when it exists",
 		},
@@ -655,7 +655,7 @@ func Test_Defaults(t *testing.T) {
 			map[string]interface{}{},
 			&stepRun{
 				atLeastVersion2: true,
-				ui:              packer.TestUi(t),
+				ui:              packersdk.TestUi(t),
 			},
 			[]string{"-display", "fakedisplay"},
 			"Display option should value config display",
@@ -667,7 +667,7 @@ func Test_Defaults(t *testing.T) {
 			map[string]interface{}{},
 			&stepRun{
 				atLeastVersion2: true,
-				ui:              packer.TestUi(t),
+				ui:              packersdk.TestUi(t),
 			},
 			[]string{"-display", "gtk"},
 			"Display option should default to gtk",
