+++ packer-plugin-sdk/packer/ui_mock.go
@@ -1,8 +1,20 @@
 package packer
 
 import (
+	"bytes"
 	"io"
+	"io/ioutil"
+	"testing"
 )
+
+func TestUi(t *testing.T) Ui {
+	var buf bytes.Buffer
+	return &BasicUi{
+		Reader:      &buf,
+		Writer:      ioutil.Discard,
+		ErrorWriter: ioutil.Discard,
+	}
+}
 
 type MockUi struct {
 	AskCalled      bool
